+++ src/pages/analytics/Analytics.js
@@ -28,15 +28,16 @@ const Analytics = ({ sectionKey }) => {
         useAnalytics()
 
     const handleStartAnalyticsTablesGeneration = () => {
-        const params = {}
-        for (const [key, { checked }] of Object.entries(checkboxes)) {
+        const params = checkboxes.reduce((obj, { key, checked }) => {
             if (checked) {
-                params[key] = true
+                obj[key] = checked
             }
-        }
+            return obj
+        }, {})
         if (lastYears !== DEFAULT_LAST_YEARS) {
             params[LAST_YEARS_INPUT_KEY] = lastYears
         }
+
         startAnalyticsTablesGeneration(params)
     }
     const handleLastYearsChange = ({ selected }) => {
@@ -57,20 +58,18 @@ const Analytics = ({ sectionKey }) => {
                 )}
                 <div className={styles.form}>
                     <div className={styles.checkboxes}>
-                        {Object.entries(checkboxes).map(
-                            ([key, { label, checked }]) => (
-                                <Checkbox
-                                    key={key}
-                                    className={styles.checkbox}
-                                    label={label}
-                                    checked={checked}
-                                    onChange={() => {
-                                        toggleCheckbox(key)
-                                    }}
-                                    disabled={loading}
-                                />
-                            )
-                        )}
+                        {checkboxes.map(({ key, label, checked }) => (
+                            <Checkbox
+                                key={key}
+                                className={styles.checkbox}
+                                label={label}
+                                checked={checked}
+                                onChange={() => {
+                                    toggleCheckbox(key)
+                                }}
+                                disabled={loading}
+                            />
+                        ))}
                     </div>
                     <SingleSelectField
                         className={styles.lastYearsSelect}