+++ src/pages/analytics/use-checkboxes.js
@@ -2,24 +2,19 @@ import { useState } from 'react'
 import { analyticsCheckboxes } from './analytics.conf.js'
 
 export const useCheckboxes = () => {
-    const [checkboxes, setCheckboxes] = useState(() => {
-        const checkboxes = {}
-        analyticsCheckboxes.forEach((checkbox) => {
-            checkboxes[checkbox.key] = {
-                checked: false,
-                label: checkbox.label,
-            }
-        })
-        return checkboxes
-    })
+    const [checkboxes, setCheckboxes] = useState(analyticsCheckboxes)
+
     const toggleCheckbox = (key) => {
-        setCheckboxes({
-            ...checkboxes,
-            [key]: {
-                ...checkboxes[key],
-                checked: !checkboxes[key].checked,
-            },
-        })
+        setCheckboxes(
+            checkboxes.slice().map((c) =>
+                c.key !== key
+                    ? c
+                    : {
+                          ...c,
+                          checked: !c.checked,
+                      }
+            )
+        )
     }
 
     return {